+++ src/index.css
@@ -111,12 +111,7 @@ button:hover {
   width: 1px;
 }
 
-/***** Tablet *****/
-
-@media (min-width: 768px) and (max-width: 1023px) {
-}
-
-/***** Desktop *****/
+/***** Tablet and Desktop *****/
 @media (min-width: 768px) {
   legend {
     text-align: center;
